+++ src/frformat/formats/code_commune_insee_frozenset.py
