+++ src/frformat/formats/code_postal.py
@@ -1,5 +1,5 @@
 from frformat import set_format
-from frformat.geo.code_postal_frozenset import CODES_POSTAUX
+from frformat.formats.code_postal_frozenset import CODES_POSTAUX
 
 name = "Code postal"
 description = "Vérifie que le code postal est bien un code postal français"
