+++ src/frformat/formats/code_postal_frozenset.py
