+++ src/frformat/formats/coordonnees_gps_francaises.py
