+++ src/frformat/formats/latitude_l93.py
