+++ src/frformat/formats/numero_departement.py
